pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbench"
version = "0.1.0"
description = "Process-based benchmark harness for evaluating LLM memory systems and agentic tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "datasets>=4.4.1",
    "huggingface-hub>=0.20.0",
    "genai-prices>=0.0.36",
    "openai>=1.0.0",
    "pydantic>=2.9",
    "pydantic-ai>=1.12.0",
    "numpy>=1.26.0",
    "smolagents>=1.22.0",
    "ddgs>=9.9.0",
    "wikipedia-api>=0.8.1",
    "requests>=2.32.5",
    "markdownify>=1.2.0",
    "mem0ai>=1.0.0",
    "memorisdk>=2.3.2",
    "ruff>=0.14.4",
    "gitpython>=3.1.40",
    "tiktoken>=0.8.0",
    "portpicker>=1.5.0",
    "chromadb>=1.3.4",
    "psutil>=7.1.3",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.15.1",
]

[project.scripts]
agentbench = "agentbench.cli:cli_main"

[project.optional-dependencies]
mem0 = ["mem0ai>=0.1.0"]
memorisdk = ["memorisdk>=2.3.2"]
smolagents = ["smolagents[toolkit]>=1.0.0", "wikipedia-api>=0.7.0"]
all = ["mem0ai>=0.1.0", "memorisdk>=2.3.2", "smolagents[toolkit]>=1.0.0", "wikipedia-api>=0.7.0"]
dev = ["ruff>=0.14.4", "pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentbench*"]
exclude = ["data*", "runs*", "tests*", "memharness*"]
